backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
paho-mqtt==2.1.0
pydantic==2.9.0
python-multipart==0.0.9

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-httpx==0.30.0
